refactor: seperate scripts for building and push
