+++ .github/workflows/autogenerated-deploy-count-files.yml
@@ -35,7 +35,7 @@ jobs:
       - run: npm run test count-files
         name: test problem
         working-directory: ./packages/problems
-      - run: npm run build count-files
+      - run: npm run push count-files
         name: push problem
         working-directory: ./packages/problems
         env:
