+++ .github/workflows/autogenerated-deploy-filter-logs.yml
@@ -35,7 +35,7 @@ jobs:
       - run: npm run test filter-logs
         name: test problem
         working-directory: ./packages/problems
-      - run: npm run build filter-logs
+      - run: npm run push filter-logs
         name: push problem
         working-directory: ./packages/problems
         env:
