+++ .github/workflows/autogenerated-deploy-grep-version-2.yml
@@ -35,7 +35,7 @@ jobs:
       - run: npm run test grep-version-2
         name: test problem
         working-directory: ./packages/problems
-      - run: npm run build grep-version-2
+      - run: npm run push grep-version-2
         name: push problem
         working-directory: ./packages/problems
         env:
