+++ .github/workflows/autogenerated-deploy-post-it.yml
@@ -35,7 +35,7 @@ jobs:
       - run: npm run test post-it
         name: test problem
         working-directory: ./packages/problems
-      - run: npm run build post-it
+      - run: npm run push post-it
         name: push problem
         working-directory: ./packages/problems
         env:
