+++ .github/workflows/autogenerated-deploy-rm-that-file.yml
@@ -35,7 +35,7 @@ jobs:
       - run: npm run test rm-that-file
         name: test problem
         working-directory: ./packages/problems
-      - run: npm run build rm-that-file
+      - run: npm run push rm-that-file
         name: push problem
         working-directory: ./packages/problems
         env:
