+++ packages/problems/scripts/build.ts
@@ -20,11 +20,6 @@ async function dockerBuild({ tag, dir }: { tag: string; dir: string }) {
   await $`docker build -t ${tag} ${dir}`
 }
 
-async function dockerPush(tag: string) {
-  if (env.DOCKER_REGISTRY === "") return
-  await $`docker push ${env.DOCKER_REGISTRY}${tag}`
-}
-
 async function init() {
   await mkdir(`${WORKING_DIR}/images/easyshell-base`, {
     recursive: true,
@@ -161,23 +156,6 @@ ENTRYPOINT ["/entrypoint"]
   return tasks
 }
 
-async function pushProblemTasks(problem: string): Promise<Array<Task>> {
-  const tasks: Array<Task> = []
-  const info = await getProblemInfo(problem)
-  for (const testcase of info.testcases) {
-    const tag = `easyshell-${problem}-${testcase.id}`
-    tasks.push({
-      name: `push-${tag}`,
-      callable: async () => {
-        if (env.DOCKER_REGISTRY === "") return "skipped"
-        await dockerPush(tag)
-        return "done"
-      },
-    })
-  }
-  return tasks
-}
-
 async function main() {
   const args = process.argv.slice(2)
   if (args.length === 0) {
@@ -193,15 +171,12 @@ async function main() {
   const arg = args[0]!
 
   const build_tasks: Array<Task> = []
-  const push_tasks: Array<Task> = []
 
   const problems = await getProblems()
   if (arg === "all") {
     await init()
     for (const problem of problems)
       build_tasks.push(...(await buildProblemTasks(problem)))
-    for (const problem of problems)
-      push_tasks.push(...(await pushProblemTasks(problem)))
   } else {
     if (!problems.includes(arg)) {
       console.error(`Problem not found: ${arg}`)
@@ -209,7 +184,6 @@ async function main() {
     }
     await init()
     build_tasks.push(...(await buildProblemTasks(arg)))
-    push_tasks.push(...(await pushProblemTasks(arg)))
   }
 
   console.log()
@@ -220,18 +194,6 @@ async function main() {
   await RunParallelStuff({
     tasks: build_tasks,
     parallel_limit: env.PARALLEL_LIMIT_BUILD,
-  })
-  console.log(
-    "================================================================================",
-  )
-  console.log()
-  console.log(
-    "=================================== Pushing ====================================",
-  )
-  console.log()
-  await RunParallelStuff({
-    tasks: push_tasks,
-    parallel_limit: env.PARALLEL_LIMIT_PUSH,
   })
   console.log(
     "================================================================================",
